+++ CHANGES.rst
@@ -68,7 +68,7 @@ charge_migration
 - Step was renamed from undersampling_migration. Changed default signal threshold,
   added efficient routine to flag neighborhood pixels, added new unit test,
   improved earlier unit tests, updated docs. [#7825]
-  
+
 cube_build
 ----------
 
@@ -193,7 +193,15 @@ residual_fringe
 
 - Use scipy.interpolate.BSpline instead of astropy.modeling.Spline1D in
   residual_fringe fitting utils [#7764]
-  
+
+
+set_telescope_pointing
+----------------------
+
+- Update the WCS calculations for GUIDING modes to match the actual operation
+  of the different FGS guiding modes. Previously, the algorithm used was the
+  same for all modes. [#7889]
+
 undersampling_correction
 ------------------------
 
@@ -202,7 +210,7 @@ undersampling_correction
 
 - Removed directories for undersampling_correction step, as the step has been
   renamed charge_migration. [#7850]
-  
+
 
 1.11.4 (2023-08-14)
 ===================